==> 1_climate_data_netcdf_to_csv.ipynb <==
#import libraries
import glob
import numpy as np
import pandas as pd
import xarray as xr
import netCDF4

#create list of days from netcdf names
all_days_pr = []
for file in glob.glob('./climate_data/Precipitation/rr_mrg_*.nc'):
    all_days_pr.append(file[23:31])

#concatenate all files of pr
pr = xr.concat([xr.open_dataset(f) for f in glob.glob('./climate_data/Precipitation/rr_mrg_*.nc')], pd.Index(all_days_pr, name="time"))

#extract to netcdf
pr.to_netcdf('./climate_data/pr.nc')

#for tmax
all_days_tmax = []
for file in glob.glob('./climate_data/Tmax/tmax_mrg_*.nc'):
    all_days_tmax.append(file[16:24])
all_days_tmax

#concatenate all files of tmax
tmax = xr.concat([xr.open_dataset(f) for f in glob.glob('./climate_data/Tmax/tmax_mrg_*.nc')], pd.Index(all_days_tmax, name="time"))

#extract to netcdf
tmax.to_netcdf('./climate_data/tmax.nc')

#for tmin
all_days_tmin = []
for file in glob.glob('./climate_data/Tmin/tmin_mrg_*.nc'):
    all_days_tmin.append(file[16:24])
all_days_tmin

#concatenate all files of tmin
tmin = xr.concat([xr.open_dataset(f) for f in glob.glob('./climate_data/Tmin/tmin_mrg_*.nc')], pd.Index(all_days_tmin, name="time"))

#extract to netcdf
tmin.to_netcdf('./climate_data/tmin.nc')

#calculate mean of variable in area
pr_fieldmean = pr.groupby("time").mean(["Lat", "Lon"])
tmax_fieldmean = tmax.groupby("time").mean(["Lat", "Lon"])
tmin_fieldmean = tmin.groupby("time").mean(["Lat", "Lon"])

#extract to csv
pr_fieldmean = pr_fieldmean['precip']  
pr_fieldmean.to_pandas().to_csv('./climate_data/pr_fieldmean.csv')

tmax_fieldmean = tmax_fieldmean['temp']  
tmax_fieldmean.to_pandas().to_csv('./climate_data/tmax_fieldmean.csv')

tmin_fieldmean = tmin_fieldmean['temp']  
tmin_fieldmean.to_pandas().to_csv('./climate_data/tmin_fieldmean.csv')

#check netcdf files
#open data with netcdf4 module
nc = netCDF4.Dataset('./climate_data/pr.nc')
print(nc.dimensions)
