==> 2_hydro_data_xlsx_to_csv.ipynb <==
#import libraries
import xlrd
import openpyxl
import numpy as np
import pandas as pd

#label of data
print("""Name of selected of data:
=====================================
stations_yearstart_yearend_source
=====================================
Stations:
ansongo: a
kandadji: k
niamey: n
=====================================
Years:
1981
2000
2010
2020
=====================================
Source:
M. Bernard (AGHRYMET): be
M. Hamatan (AGHRYMET): ha
M. Bachir (ABN): ba
=====================================""")

## Load hydro data

#loaded data
#rename headers
#drop useless rows at head (optimal)
#drop useless rows at tail (optimal)
#replace -9999.0 by NaN (optimal)
#split columns of dataframe (optimal)
#set date in datetime
#delete double dates (optimal)
#set index and drop index column (optimal)
#view dataframe

### Bernard AGRHYMET

akn_1981_2016_be = pd.read_excel('./hydro_data/Ansongo_Kandadji_Niamey_Debits_journaliers_real_bernard_agrhymet.xlsx', engine='openpyxl')

akn_1981_2016_be = akn_1981_2016_be.drop(akn_1981_2016_be.index[13051:13088])
akn_1981_2016_be = akn_1981_2016_be.replace(-9999.0, np.nan)
akn_1981_2016_be = akn_1981_2016_be.rename(columns={'DATE':'Date'})

a_1981_2016_be = akn_1981_2016_be[['Date', 'Ansongo']]
k_1981_2016_be = akn_1981_2016_be[['Date', 'Kandadji']]
n_1981_2016_be = akn_1981_2016_be[['Date', 'Niamey']]

a_1981_2016_be['Date'] = pd.to_datetime(akn_1981_2016_be['Date'], infer_datetime_format=True) 
k_1981_2016_be['Date'] = pd.to_datetime(akn_1981_2016_be['Date'], infer_datetime_format=True)  
n_1981_2016_be['Date'] = pd.to_datetime(akn_1981_2016_be['Date'], infer_datetime_format=True) 
#akn_1981_2016_be = pd.to_datetime(akn_1981_2016_be.pop('Date'), infer_datetime_format=True)

#akn_1981_2016_be
#a_1981_2016_be
#k_1981_2016_be
#n_1981_2016_be

#save data to csv
#a_1981_2016_be.to_csv('./hydro_data/csv/a_1981_2016_be.csv', index=False)
#k_1981_2016_be.to_csv('./hydro_data/csv/k_1981_2016_be.csv', index=False)
#n_1981_2016_be.to_csv('./hydro_data/csv/n_1981_2016_be.csv', index=False)
#akn_1981_2016_be.to_csv('./hydro_data/csv/akn_1981_2016_be.csv', index=False)

### Hamatan AGRHYMET

a_1981_2010_ha = pd.read_excel('./hydro_data/Qj_Ansongo_1981-2010_real_hamatan_agrhymet.xlsx', engine='openpyxl')

a_1981_2010_ha = a_1981_2010_ha.rename(columns={'Data':'Ansongo'})
a_1981_2010_ha['Date'] = pd.to_datetime(a_1981_2010_ha['Date'], infer_datetime_format=True)

a_1981_2010_ha

#save data to csv
#a_1981_2010_ha.to_csv('./hydro_data/csv/a_1981_2010_ha.csv', index=False)

k_1981_2010_ha = pd.read_excel('./hydro_data/Qj_Kandadji_1981-2010_real_hamatan_agrhymet.xlsx', engine='openpyxl')

k_1981_2010_ha = k_1981_2010_ha.rename(columns={'date':'Date', 'Qj':'Kandadji'})
k_1981_2010_ha['Date'] = pd.to_datetime(k_1981_2010_ha['Date'], infer_datetime_format=True)

k_1981_2010_ha

#save data to csv
#k_1981_2010_ha.to_csv('./hydro_data/csv/k_1981_2010_ha.csv', index=False)

n_1981_2011_ha = pd.read_excel('./hydro_data/Qj_Niamey_1981-2010_real_hamatan_agrhymet_modif.xlsx', engine='openpyxl')

n_1981_2011_ha['Date'] = pd.to_datetime(n_1981_2011_ha['Date'], infer_datetime_format=True)

n_1981_2011_ha

#save data to csv
#n_1981_2010_ha.to_csv('./hydro_data/csv/n_1981_2010_ha.csv', index=False)

### Bachir AGRHYMET

a_1980_2010_ba = pd.read_excel('./hydro_data/Niamey-Ansongo_1980-2010_bachir_abn.xlsx', sheet_name='Ansongo_1980-2010', engine='openpyxl')

a_1980_2010_ba = a_1980_2010_ba.rename(columns={"Débit (m3/s)":'Ansongo'})

a_1980_2010_ba = a_1980_2010_ba.drop([0])
a_1980_2010_ba = a_1980_2010_ba.drop(a_1980_2010_ba.index[10604:22689])
a_1980_2010_ba = a_1980_2010_ba.reset_index().drop('index', axis=1)

a_1980_2010_ba['Date'] = pd.to_datetime(a_1980_2010_ba['Date'], infer_datetime_format=True)

a_1980_2010_ba

#save data to csv
#a_1980_2010_ba.to_csv('./hydro_data/csv/a_1980_2010_ba.csv', index=False)

a_2000_2020_ba = pd.read_excel('./hydro_data/Débits_Ansongo_2000-2020_bachir_abn.xlsx', engine='openpyxl')

a_2000_2020_ba = a_2000_2020_ba.rename(columns={"Débits à la station d'Ansongo":'Date', "Unnamed: 1":'Ansongo'})

a_2000_2020_ba = a_2000_2020_ba.drop([0,1])
a_2000_2020_ba = a_2000_2020_ba.drop(a_2000_2020_ba.index[7276:7550])
a_2000_2020_ba = a_2000_2020_ba.reset_index().drop('index', axis=1)

a_2000_2020_ba['Date'] = pd.to_datetime(a_2000_2020_ba['Date'], infer_datetime_format=True)

a_2000_2020_ba

#save data to csv
#a_2000_2020_ba.to_csv('./hydro_data/csv/a_2000_2020_ba.csv', index=False)

k_1980_2010_ba = pd.read_excel('./hydro_data/Débit_Kandadji_1980-2010_bachir_abn.xlsx', 
                               names=['Date', 'Kandadji'], engine='openpyxl')

k_1980_2010_ba = k_1980_2010_ba.drop(k_1980_2010_ba.index[9903:11338])

k_1980_2010_ba['Date'] = pd.to_datetime(k_1980_2010_ba['Date'], infer_datetime_format=True)

k_1980_2010_ba

#save data to csv
#k_1980_2010_ba.to_csv('./hydro_data/csv/k_1980_2010_ba.csv', index=False)

k_2000_2020_ba = pd.read_excel('./hydro_data/Niamey_Kandadji_2000-2020_bachir_abn.xlsx', sheet_name='Kandadji', engine='openpyxl')

k_2000_2020_ba = k_2000_2020_ba.rename(columns={"Débits (m3/s)":'Kandadji'})

k_2000_2020_ba = k_2000_2020_ba.drop([0])
k_2000_2020_ba = k_2000_2020_ba.drop(k_2000_2020_ba.index[6615:14241])
k_2000_2020_ba = k_2000_2020_ba.reset_index().drop('index', axis=1)

k_2000_2020_ba = k_2000_2020_ba.drop(['Unnamed: 2', 'Unnamed: 3', 'Unnamed: 4', 'Unnamed: 5', 'Unnamed: 6',
                                      'Unnamed: 7', 'Unnamed: 8', 'Unnamed: 9'], axis='columns')

k_2000_2020_ba['Date'] = pd.to_datetime(k_2000_2020_ba['Date'], infer_datetime_format=True)

k_2000_2020_ba

#save data to csv
#k_2000_2020_ba.to_csv('./hydro_data/csv/k_2000_2020_ba.csv', index=False)

n_1980_2010_ba = pd.read_excel('./hydro_data/Niamey-Ansongo_1980-2010_bachir_abn.xlsx', sheet_name="Niamey 1980-2010", engine='openpyxl')

n_1980_2010_ba = n_1980_2010_ba.rename(columns={"Débit (m3/s)":'Niamey'})

n_1980_2010_ba = n_1980_2010_ba.drop([0])
n_1980_2010_ba = n_1980_2010_ba.reset_index().drop('index', axis=1)

n_1980_2010_ba['Date'] = pd.to_datetime(n_1980_2010_ba['Date'], infer_datetime_format=True)

n_1980_2010_ba

#save data to csv
#n_1980_2010_ba.to_csv('./hydro_data/csv/n_1980_2010_ba.csv', index=False)

n_2002_2020_ba = pd.read_excel('./hydro_data/Niamey_Kandadji_2000-2020_bachir_abn.xlsx', sheet_name='Niamey', engine='openpyxl')

n_2002_2020_ba = n_2002_2020_ba.rename(columns={"Débits (m3/s)":'Niamey'})

n_2002_2020_ba = n_2002_2020_ba.drop([0])
n_2002_2020_ba = n_2002_2020_ba.drop(n_2002_2020_ba.index[6904:6931])
n_2002_2020_ba = n_2002_2020_ba.reset_index().drop('index', axis=1)

n_2002_2020_ba['Date'] = pd.to_datetime(n_2002_2020_ba['Date'], infer_datetime_format=True)

n_2002_2020_ba

#save data to csv
#n_2002_2020_ba.to_csv('./hydro_data/csv/n_2002_2020_ba.csv', index=False)
