==> 4_merge_climate_and_hydro_data.ipynb <==
#import libraries
import pandas as pd

#set periode of study
start_date = '1981-06-01'
end_date = '2010-12-31'

#load climate data
pr = pd.read_csv('./climate_data/pr_fieldmean.csv')
tmax = pd.read_csv('./climate_data/tmax_fieldmean.csv')
tmin = pd.read_csv('./climate_data/tmin_fieldmean.csv')

#rename colums
pr = pr.rename(columns={'time':'Date', '0': 'pr'})
tmax = tmax.rename(columns={'time':'Date', '0': 'tmax'})
tmin = tmin.rename(columns={'time':'Date', '0': 'tmin'})

#merge pr, tmax and tmin by date column
climate_data = pr.merge(tmax, on='Date').merge(tmin, on='Date')

#select time periode between 19810601 and 20101231
climate_data = climate_data.loc[(climate_data['Date'] >= 19810601) & (climate_data['Date'] <= 20101231)]

#convert time to datetime
climate_data['Date'] = pd.to_datetime(climate_data['Date'], format='%Y%m%d')

#reset index of climate data
climate_data = climate_data.reset_index().drop('index', axis=1)

#load hydro data
ansongo = pd.read_csv('./hydro_data/interpolate/ansongo.csv')
kandadji = pd.read_csv('./hydro_data/interpolate/kandadji.csv')
niamey = pd.read_csv('./hydro_data/interpolate/niamey.csv')

#merge ansongo, kandadji and niamey by date column
hydro_data = ansongo.merge(kandadji, on='Date').merge(niamey, on='Date')

#define datetime of hydro data
hydro_data['Date'] = pd.to_datetime(hydro_data['Date'], infer_datetime_format=True)

#merge climate data and hydro by time column
model_data = climate_data.merge(hydro_data, on='Date')

model_data

#save data to csv
model_data.to_csv('./model_data/model_data_interpolated.csv', index=False)
